Moved Item classes to their own file
